refactor: extract material loader mixin from combo-box styles
