+++ packages/vaadin-material-styles/mixins/loader.d.ts
@@ -1,0 +1,3 @@
+import type { CSSResult } from 'lit';
+
+export const loader: CSSResult;
